+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @selvajs/ui
 
+## 2.0.7
+
+### Patch Changes
+
+- @selvajs/schemas@2.0.7
+
+## 2.0.6
+
+### Patch Changes
+
+- @selvajs/schemas@2.0.6
+
+## 2.0.5
+
+### Patch Changes
+
+- @selvajs/schemas@2.0.5
+
 ## 2.0.4
 
 ### Patch Changes